refactor: regroup Sales/Fulfillment/Account under shared AccountContainer layout
