+++ client/src/routes.tsx
@@ -159,18 +159,43 @@ const routes: RouteObject[] = [
         },
       },
       {
-        path: "account",
         async lazy() {
-          const { default: Component } =
+          const { default: AccountContainer } =
             await import("components/Account/AccountContainer");
           return {
             Component: () => (
               <AuthWrapper>
-                <Component />
+                <AccountContainer />
               </AuthWrapper>
             ),
           };
         },
+        children: [
+          {
+            path: "account",
+            async lazy() {
+              const { default: Component } =
+                await import("components/Account/AccountSettings");
+              return { Component };
+            },
+          },
+          {
+            path: "fulfillment",
+            async lazy() {
+              const { default: Component } =
+                await import("components/FulFillment/Fulfillment");
+              return { Component };
+            },
+          },
+          {
+            path: "sales",
+            async lazy() {
+              const { default: Component } =
+                await import("components/Sales/Sales");
+              return { Component };
+            },
+          },
+        ],
       },
       {
         path: "profile",
@@ -620,33 +645,6 @@ const routes: RouteObject[] = [
             },
           },
         ],
-      },
-      {
-        path: "fulfillment",
-        async lazy() {
-          const { default: Fulfillment } =
-            await import("components/FulFillment/Fulfillment");
-          return {
-            Component: () => (
-              <AuthWrapper>
-                <Fulfillment />
-              </AuthWrapper>
-            ),
-          };
-        },
-      },
-      {
-        path: "sales",
-        async lazy() {
-          const { default: Sales } = await import("components/Sales/Sales");
-          return {
-            Component: () => (
-              <AuthWrapper>
-                <Sales />
-              </AuthWrapper>
-            ),
-          };
-        },
       },
       {
         path: "releases",
